refactor: consolidate repository implementations into infrastructure subdirectory
